pyproject.toml
[tool.poetry]
name = "programming-quiz"
version = "0.1.0"
description = "Projeto referente ao P.I do primeiro semestre de 2023"
authors = ["do-ald533 <douglas.aldred533@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poe.tasks]
dev = { cmd= "reloadium run src/main.py" }
start = { script = "main:main"}

[tool.poetry.dependencies]
python = "^3.10"
pygame = "^2.2.0"
sqlalchemy = "^2.0.5.post1"
cerberus = "^1.3.4"
python-dotenv = "^1.0.0"
pymysql = "^1.0.3"
cryptography = "^40.0.2"


[tool.poetry.group.test.dependencies]
pytest = "^7.2.2"

[tool.poetry.group.dev.dependencies]
reloadium = "^1.0.0"
poethepoet = "^0.18.1"
loguru = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
